PYIC-4110: Simplify HTTP server code
