+++ libs/pact-test-helpers/build.gradle
@@ -15,9 +15,7 @@ dependencies {
 			"org.mockito:mockito-junit-jupiter:5.10.0",
 			"com.nimbusds:oauth2-oidc-sdk:$rootProject.ext.dependencyVersions.nimbusdsOauth2OidcSdk",
 			"au.com.dius.pact.consumer:junit5:4.6.5",
-			'com.fasterxml.jackson.core:jackson-databind:2.15.3',
-			"org.apache.logging.log4j:log4j-api:$rootProject.ext.dependencyVersions.log4j",
-			"org.apache.logging.log4j:log4j-core:$rootProject.ext.dependencyVersions.log4j"
+			'com.fasterxml.jackson.core:jackson-databind:2.15.3'
 
 	testImplementation "org.junit.jupiter:junit-jupiter:5.10.0",
 			"org.mockito:mockito-junit-jupiter:5.10.0",
